+++ src/main/java/com/fastcampus/housebatch/core/service/FakeSendService.java
@@ -1,4 +1,4 @@
-package com.fastcampus.housebatch.adapter;
+package com.fastcampus.housebatch.core.service;
 
 import org.springframework.stereotype.Service;
