+++ src/main/java/com/fastcampus/housebatch/core/service/SendService.java
@@ -1,4 +1,4 @@
-package com.fastcampus.housebatch.adapter;
+package com.fastcampus.housebatch.core.service;
 
 public interface SendService {
 	void send(String email, String message);
